requirements.txt
Flask==2.0.3
flask-mongoengine==1.0.0
mongoengine==0.24.2
pymongo==4.1.1
Flask-CORS==3.0.10
pandas==1.5.3
numpy==1.24.2
scikit-learn==1.2.2
werkzeug==2.0.3
python-dotenv==1.0.0
openpyxl==3.1.2